Сообщения для пользователя и обработка его ввода формируется в наследниках абстрактного класса Step. Выбор конкретной реализации Step по ключу StepSelector из мапы StepHolder.steps. Сама мапа инициализируется при старте в TestDataInit.
