+++ src/main/java/spring/app/core/StepSelector.java
@@ -1,0 +1,9 @@
+package spring.app.core;
+
+public enum StepSelector {
+    Start,
+    UserMenu,
+    AdminMenu,
+    AdminAddUser,
+    AdminRemoveUser
+}
