+++ src/main/java/spring/app/core/steps/UserMenu.java
@@ -1,0 +1,34 @@
+package spring.app.core.steps;
+
+import org.springframework.stereotype.Component;
+import spring.app.core.BotContext;
+import spring.app.exceptions.ProcessInputException;
+import spring.app.util.Keyboards;
+
+@Component
+public class UserMenu extends Step {
+    private String text;
+    private String keyboard;
+
+    @Override
+    public void enter(BotContext context) {
+        // TODO проверку из ТЗ
+        text = "Пользователь с таким vk id не найден в базе. Обратитесь к Герману Севостьянову или Станиславу Сорокину";
+        keyboard = Keyboards.defaultKeyboard;
+    }
+
+    @Override
+    public void processInput(BotContext context) throws ProcessInputException {
+
+    }
+
+    @Override
+    public String getText() {
+        return text;
+    }
+
+    @Override
+    public String getKeyboard() {
+        return keyboard;
+    }
+}
